Draft: split Path Array and viewprovider Array from Draft.Py
Draft: fix typo in PathArray
